feat: update Taskfile.yaml to uninstall pre-commit before creating index and install it afterwards
This commit updates the Taskfile.yaml to include the command 'pre-commit uninstall' before creating the index file and pushes it to the default GitHub pages branch. It also includes the command 'pre-commit install' after creating the index file. This ensures that the pre-commit hooks are uninstalled and then reinstalled properly during the index creation process.