feat: Add contributing guidelines and tools
Add a contributing section to the README.md.gotmpl file, providing guidelines for adding features or bugfixes. Also, include instructions on opening issues and providing unit tests.

Additionally, mention the use of Taskfile (https://taskfile.dev/) for convenient shortcuts to useful commands.

In order to configure the environment, users need to copy the ".env-template" file as ".env" and customize it according to their needs.