feat: Add CONTRIBUTING.md file
Add CONTRIBUTING.md file to provide guidelines for contributing to the Airsonic-Advanced Helm Chart. The file includes sections on Code of Conduct, Asking Questions, Bug Reporting, Suggesting Enhancements, First Code Contribution, Improving Documentation, Styleguides, and Joining the Project Team. The content of the file is based on the contributing-gen template.