feat: Add CODE_OF_CONDUCT.md
Add the CODE_OF_CONDUCT.md file to the Airsonic-Advanced Helm Chart repository. The file includes the project's code of conduct, which outlines the standards of behavior expected from contributors and maintainers. The code of conduct promotes a positive and inclusive environment, free from harassment and discrimination. It provides guidelines for acceptable behavior, as well as procedures for reporting and addressing instances of unacceptable behavior. The code of conduct is adapted from the Contributor Covenant version 1.4 and 2.0.